[REF] move account_check_printing mods to l10n_latam_manual_check
