+++ addons/account_check_printing/__manifest__.py
@@ -16,13 +16,11 @@ The check settings are located in the accounting journals configuration page.
         'security/ir.model.access.csv',
         'data/account_check_printing_data.xml',
         'views/account_journal_views.xml',
-        'views/account_checkbook_view.xml',
         'views/account_move_views.xml',
         'views/account_payment_views.xml',
         'views/res_config_settings_views.xml',
         'views/res_partner_views.xml',
-        'wizard/account_payment_register_views.xml',
-        'wizard/print_prenumbered_checks_views.xml',
+        'wizard/print_prenumbered_checks_views.xml'
     ],
     'installable': True,
     'auto_install': False,
