+++ addons/account_check_printing/models/__init__.py
@@ -7,4 +7,3 @@ from . import account_payment
 from . import res_company
 from . import res_config_settings
 from . import res_partner
-from . import account_checkbook
