+++ addons/account_check_printing/models/account_journal.py
@@ -30,8 +30,6 @@ class AccountJournal(models.Model):
         compute='_compute_check_printing_payment_method_selected',
         help="Technical feature used to know whether check printing was enabled as payment method.",
     )
-    use_checkbooks = fields.Boolean()
-    checkbook_ids = fields.One2many('account.checkbook', 'journal_id', 'Checkbooks', context={'active_test': False},)
 
     @api.depends('check_manual_sequencing')
     def _compute_check_next_number(self):
