+++ addons/account_check_printing/views/account_journal_views.xml
@@ -27,23 +27,16 @@
             <field name="inherit_id" ref="account.view_account_journal_form" />
             <field name="arch" type="xml">
                 <xpath expr="//page[@name='journal_entries']//group[@name='outgoing_payment']" position="after">
-                    <group string="Checks"
+                    <group string="Check Printing"
                            attrs="{'invisible': ['|', ('type', '!=', 'bank'), ('check_printing_payment_method_selected', '=', False)]}">
                         <field name="check_printing_payment_method_selected" invisible="1"/>
                         <field name="check_sequence_id" invisible="1"/>
-                        <field name="use_checkbooks"/>
-                        <field name="check_manual_sequencing" attrs="{'invisible': [('use_checkbooks', '=', True)]}"/>
-                        <!-- <label for="checkbook_ids" attrs="{'invisible': [('use_checkbooks', '=', False)]}" colspan="2"/> -->
-                        <field name="checkbook_ids" attrs="{'invisible': [('use_checkbooks', '=', False)]}" colspan="2" nolabel="1"/>
+                        <field name="check_manual_sequencing"
+                        />
                         <field name="check_next_number"
                                attrs="{'invisible': [('check_manual_sequencing', '=', False)]}"/>
                     </group>
                 </xpath>
-        <!-- <field name="arch" type="xml">
-            <page name="bank_account">
-                <field name="show_checkbooks" invisible="1"/>
-            </page>
-        </field> -->
             </field>
         </record>
 </odoo>
