+++ addons/account_check_printing/views/account_payment_views.xml
@@ -6,8 +6,7 @@
             <field name="inherit_id" ref="account.view_account_payment_form" />
             <field name="arch" type="xml">
                 <xpath expr="//button[@name='action_post']" position="before">
-                    <button name="print_checks" class="oe_highlight" attrs="{'invisible': ['|', '|', '|', ('payment_method_code', '!=', 'check_printing'), ('state', '!=', 'posted'), ('is_move_sent', '=', True), ('check_printing_type', '=', 'no_print')]}" string="Print Check" type="object"/>
-                    <button name="action_mark_sent" class="oe_highlight" attrs="{'invisible': ['|', '|', '|', ('payment_method_code', '!=', 'check_printing'), ('state', '!=', 'posted'), ('is_move_sent', '=', True), ('check_printing_type', '!=', 'no_print')]}" string="Mark Sent" type="object"/>
+                    <button name="print_checks" class="oe_highlight" attrs="{'invisible': ['|', '|', ('payment_method_code', '!=', 'check_printing'), ('state', '!=', 'posted'), ('is_move_sent', '=', True)]}" string="Print Check" type="object"/>
                     <button name="action_unmark_sent" attrs="{'invisible': ['|', ('payment_method_code', '!=', 'check_printing'), ('is_move_sent', '=', False)]}" string="Unmark Sent" type="object"/>
                     <button name="action_void_check" attrs="{'invisible': ['|', '|', ('payment_method_code', '!=', 'check_printing'), ('state', '!=', 'posted'), ('is_move_sent', '=', False)]}" string="Void Check" type="object"/>
                 </xpath>
@@ -15,13 +14,8 @@
                     <field name="check_amount_in_words" attrs="{'invisible': [('payment_method_code', '!=', 'check_printing')]}" groups="base.group_no_one"/>
                 </xpath>
                 <xpath expr="//field[@name='payment_method_id']" position="after">
-                    <field name="use_checkbooks" invisible="1"/>
-                    <field name="checkbook_type" invisible="1"/>
-                    <field name="checkbook_id" options="{'no_create': True, 'no_open': True}" attrs="{'invisible': ['|', ('payment_method_code', '!=', 'check_printing'), ('use_checkbooks', '=', False)], 'required': [('payment_method_code', '=', 'check_printing'), ('use_checkbooks', '=', True)]}" domain="[('journal_id', '=', journal_id)]"/>
-                    <field name="check_printing_type" invisible="1"/>
                     <field name="check_manual_sequencing" invisible="1"/>
-                    <field name="check_number" attrs="{'invisible': ['|', ('payment_method_code', '!=', 'check_printing'), ('check_number', '=', False), ('check_printing_type', '!=', 'no_print')], 'readonly': ['|', ('state', '!=', 'draft'), ('check_printing_type', '!=', 'no_print')], 'required': [('check_printing_type', '=', 'no_print')]}" />
-                    <field name="check_payment_date" attrs="{'invisible': [('checkbook_type', '!=', 'deferred')], 'required': [('checkbook_type', '=', 'deferred')]}"/>
+                    <field name="check_number" attrs="{'invisible': ['|', ('payment_method_code', '!=', 'check_printing'), ('check_number', '=', False)]}"/>
                 </xpath>
                 <xpath expr="//field[@name='name']" position='before'>
                     <widget name="web_ribbon" title="Sent" attrs="{'invisible': [('is_move_sent', '!=', True)]}"/>
@@ -41,68 +35,5 @@
                 </xpath>
             </field>
         </record>
-
-    <!-- Own Check Views and menus -->
-
-        <record model="ir.ui.view" id="view_account_check_calendar">
-            <field name="name">account.check.calendar</field>
-            <field name="model">account.payment</field>
-            <field name="arch" type="xml">
-                <calendar
-                        mode="month"
-                        date_start="check_payment_date"
-                        color="journal_id">
-                    <field name="amount"/>
-                </calendar>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_account_check_pivot">
-            <field name="name">account.check.calendar</field>
-            <field name="model">account.payment</field>
-            <field name="arch" type="xml">
-                <pivot>
-                    <field name="check_payment_date" type="row" interval="month"/>
-                    <field name="check_payment_date" type="row" interval="week"/>
-                <field name="amount" type="measure"/>
-                </pivot>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_account_check_tree">
-            <field name="name">account.check.tree</field>
-            <field name="model">account.payment</field>
-            <field name="priority">100</field>
-            <field name="mode">primary</field>
-            <field name="inherit_id" ref="account.view_account_payment_tree"/>
-            <field name="arch" type="xml">
-                <field name="payment_method_id" position="replace"/>
-                <field name="name" position="replace">
-                    <field name="check_number"/>
-                </field>
-                <field name="date" position="after">
-                    <field name="check_payment_date" string="Payment Date" optional="hide"/>
-                </field>
-                <tree>
-                    <field name="is_matched"/>
-                </tree>
-            </field>
-        </record>
-
-        <record model="ir.actions.act_window" id="action_own_check">
-            <field name="name">Own Checks</field>
-            <field name="res_model">account.payment</field>
-            <field name="view_mode">tree,form,calendar,graph,pivot</field>
-            <field name="view_id" ref="view_account_check_tree"/>
-            <field name="domain">[('payment_method_code', '=', 'check_printing')]</field>
-            <!-- <field name="context">{'search_default_state':'handed', 'default_type':'own_check'}</field> -->
-            <field name="context">{'search_default_is_matched': True}</field>
-        </record>
-
-        <menuitem
-            action="action_own_check"
-            id="menu_own_check"
-            sequence="50"
-            parent="account.menu_finance_payables"/>
 
 </odoo>
