+++ addons/l10n_latam_manual_check/__init__.py
@@ -1,0 +1,2 @@
+from . import models
+from . import wizards
