+++ addons/l10n_latam_manual_check/models/__init__.py
@@ -1,0 +1,3 @@
+from . import account_checkbook
+from . import account_payment
+from . import account_journal
