+++ addons/l10n_latam_manual_check/models/account_journal.py
@@ -1,0 +1,9 @@
+from odoo import fields, models
+
+
+class AccountJournal(models.Model):
+
+    _inherit = 'account.journal'
+
+    use_checkbooks = fields.Boolean()
+    checkbook_ids = fields.One2many('account.checkbook', 'journal_id', 'Checkbooks', context={'active_test': False},)
