+++ addons/l10n_latam_manual_check/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
+"account_checkbook_access_full","account_checkbook_access_full","model_account_checkbook","account.group_account_invoice",1,1,1,1
+"account_checkbook_access_global","account_checkbook_access_global","model_account_checkbook","base.group_user",1,1,1,1
