+++ addons/l10n_latam_manual_check/views/account_checkbook_view.xml
@@ -15,23 +15,4 @@
         </field>
     </record>
 
-<!-- 
-    <record id="view_account_checkbook_form" model="ir.ui.view">
-        <field name="name">account.checkbook.form</field>
-        <field name="model">account.checkbook</field>
-        <field name="arch" type="xml">
-            <form>
-                <sheet string="Checkbooks">
-                    <group>
-                        <field name="name" invisible="1"/>
-                        <field name="sequence_id" groups="base.group_no_one"/>
-                        <field name="next_number"/>
-                        <field name="type"/>
-                        <field name="range_to"/>
-                        <field name="numerate_on_printing"/>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record> -->
 </odoo>
