+++ addons/l10n_latam_manual_check/views/account_journal_view.xml
@@ -1,0 +1,19 @@
+<odoo>
+
+    <record id="view_account_journal_tree" model="ir.ui.view">
+        <field name="name">account.journal.tree</field>
+        <field name="model">account.journal</field>
+        <field name="inherit_id" ref="account_check_printing.view_account_journal_form_inherited"/>
+        <field name="arch" type="xml">
+            <field name="check_manual_sequencing" position="attributes">
+                <attribute name="attrs">{'invisible': [('use_checkbooks', '=', True)]}</attribute>
+            </field>
+            <field name="check_sequence_id" position="after">
+                <field name="use_checkbooks"/>
+                <!-- <label for="checkbook_ids" attrs="{'invisible': [('use_checkbooks', '=', False)]}" colspan="2"/> -->
+                <field name="checkbook_ids" attrs="{'invisible': [('use_checkbooks', '=', False)]}" colspan="2" nolabel="1"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
