+++ addons/l10n_latam_manual_check/views/account_payment_view.xml
@@ -1,0 +1,87 @@
+<odoo>
+
+    <record id="view_account_payment_tree" model="ir.ui.view">
+        <field name="name">account.payment.tree</field>
+        <field name="model">account.payment</field>
+        <field name="inherit_id" ref="account_check_printing.view_account_payment_form_inherited"/>
+        <field name="arch" type="xml">
+            <button name="print_checks" position="attributes">
+                <attribute name="attrs">{'invisible': ['|', '|', '|', ('payment_method_code', '!=', 'check_printing'), ('state', '!=', 'posted'), ('is_move_sent', '=', True), ('check_printing_type', '=', 'no_print')]}</attribute>
+            </button>
+            <button name="print_checks" position="after">
+                <button name="action_mark_sent" class="oe_highlight" attrs="{'invisible': ['|', '|', '|', ('payment_method_code', '!=', 'check_printing'), ('state', '!=', 'posted'), ('is_move_sent', '=', True), ('check_printing_type', '!=', 'no_print')]}" string="Mark Sent" type="object"/>
+            </button>
+            <xpath expr="//field[@name='payment_method_id']" position="after">
+                <field name="use_checkbooks" invisible="1"/>
+                <field name="checkbook_type" invisible="1"/>
+                <field name="checkbook_id" options="{'no_create': True, 'no_open': True}" attrs="{'invisible': ['|', ('payment_method_code', '!=', 'check_printing'), ('use_checkbooks', '=', False)], 'required': [('payment_method_code', '=', 'check_printing'), ('use_checkbooks', '=', True)]}" domain="[('journal_id', '=', journal_id)]"/>
+                <field name="check_printing_type" invisible="1"/>
+                <field name="check_number" attrs="{'invisible': ['|', ('payment_method_code', '!=', 'check_printing'), ('check_number', '=', False), ('check_printing_type', '!=', 'no_print')], 'readonly': ['|', ('state', '!=', 'draft'), ('check_printing_type', '!=', 'no_print')], 'required': [('check_printing_type', '=', 'no_print')]}" />
+                <field name="check_payment_date" attrs="{'invisible': [('checkbook_type', '!=', 'deferred')], 'required': [('checkbook_type', '=', 'deferred')]}"/>
+            </xpath>
+        </field>
+    </record>
+
+ <!-- Own Check Views and menus -->
+
+    <record model="ir.ui.view" id="view_account_check_calendar">
+        <field name="name">account.check.calendar</field>
+        <field name="model">account.payment</field>
+        <field name="arch" type="xml">
+            <calendar
+                    mode="month"
+                    date_start="check_payment_date"
+                    color="journal_id">
+                <field name="amount"/>
+            </calendar>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_account_check_pivot">
+        <field name="name">account.check.calendar</field>
+        <field name="model">account.payment</field>
+        <field name="arch" type="xml">
+            <pivot>
+                <field name="check_payment_date" type="row" interval="month"/>
+                <field name="check_payment_date" type="row" interval="week"/>
+            <field name="amount" type="measure"/>
+            </pivot>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_account_check_tree">
+        <field name="name">account.check.tree</field>
+        <field name="model">account.payment</field>
+        <field name="priority">100</field>
+        <field name="mode">primary</field>
+        <field name="inherit_id" ref="account.view_account_payment_tree"/>
+        <field name="arch" type="xml">
+            <field name="payment_method_id" position="replace"/>
+            <field name="name" position="replace">
+                <field name="check_number"/>
+            </field>
+            <field name="date" position="after">
+                <field name="check_payment_date" string="Payment Date" optional="hide"/>
+            </field>
+            <tree>
+                <field name="is_matched"/>
+            </tree>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="action_own_check">
+        <field name="name">Own Checks</field>
+        <field name="res_model">account.payment</field>
+        <field name="view_mode">tree,form,calendar,graph,pivot</field>
+        <field name="view_id" ref="view_account_check_tree"/>
+        <field name="domain">[('payment_method_code', '=', 'check_printing')]</field>
+        <!-- <field name="context">{'search_default_state':'handed', 'default_type':'own_check'}</field> -->
+        <field name="context">{'search_default_is_matched': True}</field>
+    </record>
+
+    <menuitem
+        action="action_own_check"
+        id="menu_own_check"
+        sequence="50"
+        parent="account.menu_finance_payables"/>
+</odoo>
