+++ addons/l10n_latam_manual_check/wizards/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import print_prenumbered_checks
+from . import account_payment_register
