+++ addons/l10n_latam_manual_check/wizards/account_payment_register_views.xml
